Uploaded a Demo video
This video is split into 3 parts: P1 - a demo of the game, P2 - shows how the screens and states change in the background as the game processes, P3 - shows the events received by the event processor when a pygame event enters.
PS: events in P3 are all ids, you can check out their meanings in lib.settings.